+++ translation/service.go
@@ -3,7 +3,6 @@ package translation
 import (
 	"bytes"
 	"context"
-	"github.com/xmidt-org/candlelight"
 
 	"net/http"
 
@@ -16,7 +15,7 @@ import (
 // Service represents the Webpa-Tr1d1um component that translates WDMP data into WRP
 // which is compatible with the XMiDT API.
 type Service interface {
-	SendWRP(context.Context,*wrp.Message, string) (*common.XmidtResponse, error)
+	SendWRP(context.Context, *wrp.Message, string) (*common.XmidtResponse, error)
 }
 
 // ServiceOptions defines the options needed to build a new translation WRP service.
@@ -56,7 +55,7 @@ type service struct {
 }
 
 // SendWRP sends the given wrpMsg to the XMiDT cluster and returns the response if any.
-func (w *service) SendWRP(ctx context.Context,wrpMsg *wrp.Message, authHeaderValue string) (*common.XmidtResponse, error) {
+func (w *service) SendWRP(ctx context.Context, wrpMsg *wrp.Message, authHeaderValue string) (*common.XmidtResponse, error) {
 	wrpMsg.Source = w.wrpSource
 
 	var payload []byte
@@ -82,6 +81,5 @@ func (w *service) SendWRP(ctx context.Context,wrpMsg *wrp.Message, authHeaderVal
 
 	r.Header.Set("Content-Type", wrp.Msgpack.ContentType())
 	r.Header.Set("Authorization", authHeaderValue)
-	candlelight.InjectTraceInformation(ctx,r.Header)
 	return w.transactor.Transact(r)
 }